move `numpy_add` case to `test_numpy`
