+++ test/sot/test_01_basic.py
@@ -17,7 +17,6 @@ import unittest
 from test_case_base import TestCaseBase
 
 import paddle
-from paddle.jit.sot.utils import strict_mode_guard
 
 
 def foo(x: int, y: paddle.Tensor):
@@ -27,19 +26,6 @@ def foo(x: int, y: paddle.Tensor):
 class TestBasic(TestCaseBase):
     def test_simple(self):
         self.assert_results(foo, 1, paddle.to_tensor(2))
-
-
-def numpy_add(x, y):
-    out = paddle.to_tensor(x.numpy() + y.numpy())
-    return out
-
-
-class TestNumpyAdd(TestCaseBase):
-    @strict_mode_guard(False)
-    def test_numpy_add(self):
-        x = paddle.to_tensor([2])
-        y = paddle.to_tensor([3])
-        self.assert_results(numpy_add, x, y)
 
 
 if __name__ == "__main__":
